requirements.txt
Flask==2.0.3 # Core framework
Flask-Mail==0.9.1 # Emails
Flask-SQLAlchemy==2.5.1 # Database ineractions between Flask and SQLAlchemy
SQLAlchemy==1.4.27 # ORM, Object Relational Mapper - allows for database interaction
Werkzeug==2.0.3 # WSGI, Web Server Gateway Interface - #https://testdriven.io/blog/what-is-werkzeug/
python-dotenv==0.19.0 # Reads environment variables in our .env file in key value pairs
gunicorn>=21.0.0 # Docker WSG Server